django-proj/src/requirements.txt
Django==3.2.8
mysqlclient==2.0.3
pandas==1.3.3
dask[dataframe]==2021.9.1
jellyfish==0.8.*
gunicorn==20.0.4
djangorestframework
requests==2.26.0
